changed how vendor_id and organization are mapped as inputs
Because the fields are conditionally displayed, the one not displayed
is not contained in the params that get passed back to the DB which
means it is not nulled out.   I, kind of, feel like for all of our
mapInputs, we should use the “|| null”, but it only would matter when a
field is not included in the form.